refactored fields-extractor-engine code
Summary : `getFields()` function returns a fields object which contains the
   description of the question.

`getOptions()` is now a single function which requires you to pass the
question type of the question and the element whose options are to be
fetched rather than calling a separate `getOptions_<QuestionType>()` for
each type of question.

The conditional checking has been done with switch-case for better code
quality rather than if else-if etc

The return type of getOptions() is an object itself rather than the
options array, and the remaining fields are being attached to the fields
attribute via object restructuring, no need to explicitly attach the
object attribute.